FBP-325. Improved cascade handling on element removal
