+++ xml-parser-core/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/bpmn/parser/core/BaseBpmnParser.kt
@@ -534,10 +534,10 @@ abstract class BaseBpmnParser: BpmnParser {
 
             var (attrName, attrValue) = attributeSelector?.split("=") ?: listOf(null, null)
             if (true == attrValue?.contains('@')) {
-                if (null == value && null == valueIndexInArray) { // Skip null unindexable props
+                if (null == value || null == valueIndexInArray) { // Skip null unindexable props
                     return
                 }
-                attrValue = attrValue.replace("@", valueIndexInArray!!.removeAt(0))
+                attrValue = attrValue.replace("@", valueIndexInArray.removeAt(0))
             }
 
             val child = childOf(currentNode, name, attrName, attrValue)
